django_backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
Faker==18.5.1
inflection==0.5.1
jsonschema==4.17.3
mdgen==0.1.10
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
